Move both KDFs into btclib.kdf (issue #1080)
`ansi_x9_63_kdf` leaves `btclib.ecc.dh` and `btclib.ecc`'s exports and
lands in a top-level `btclib.kdf`, with RFC 5869's `hkdf_extract`,
`hkdf_expand` and `hkdf` beside it. Source-breaking, so RELEASE_NOTES.md
gains a bullet naming both dead spellings, checked against `v2023.7.12`.

The drawer moves rather than the new KDF joining the old one where it
sat. SEC 1 does define ANSI-X9.63-KDF in section 3.6.1 for the agreement
of section 6.1, so `ecc.dh` reads naturally with one KDF in the tree;
HKDF has no such tie, and leaving the drawer under `ecc` meant either a
KDF reached through a package of signature schemes or two KDFs in two
places for the next one to choose between. A KDF is a hash construction
with no curve and no bitcoin in it, so the top level, beside `hashes`,
is the layer it belongs to.

`ecc.dh` imports the module rather than the function, so
`btclib.ecc.dh.ansi_x9_63_kdf` is not a surviving spelling either: the
break is total, which is what lets the release note say so.

Appendix A's SHA-1 vectors are run rather than trimmed, and
tests/kdf_test.py's docstring says why: the rule here is that btclib
never chooses SHA-1, and a KDF chooses no hash function at all -- `hf`
is the caller's, `ansi_x9_63_kdf` is already driven over
`[sha1, sha256, sha512]`, and running half a published vector set is
editing the authority rather than citing it.

The numeric mutation profile takes `btclib/kdf.py` and
`tests/kdf_test.py` beside `btclib/ecc/dh.py`; its enumeration was
re-derived with `cosmic-ray init` and `cr-filter-operators` rather than
adjusted by arithmetic, and is unchanged at 720 and 66 -- the code moved
modules, it did not grow.
