+++ CHANGELOG.md
@@ -2094,37 +2094,54 @@ documented at release-notes length in the first place, and are still in
 
 ### Curves, signatures and keys
 
-- **`ecc.dh.hkdf` derives keying data per RFC 5869**, extract and expand
-  (issue #1080). `hkdf_extract` concentrates the entropy of input keying
-  material into one digest under an optional salt, `hkdf_expand` stretches
-  a pseudorandom key to the requested length under optional context, and
-  `hkdf` is the composition; all three are re-exported by `btclib.ecc`.
-  It is `hmac` and `hashlib` and nothing else, which is what puts it in
-  btclib at all: issue #1066 settled BIP324's transport line as "what can
-  be built out of the standard library is in scope, what would need a
-  hand-rolled cipher is not", and HKDF-SHA256 is the one piece of that
-  key schedule on the near side of it. ChaCha20, Poly1305, the AEAD, the
-  forward-secure rekeying and the packet encoding stay out, and so do
-  BIP324's own labels, which are transport work with no consumer here.
+- **`btclib.kdf` is the key derivation functions, and it holds RFC
+  5869's HKDF beside SEC 1's `ansi_x9_63_kdf`** (issue #1080).
+  `hkdf_extract` concentrates the entropy of input keying material into
+  one digest under an optional salt, `hkdf_expand` stretches a
+  pseudorandom key to the requested length under optional context, and
+  `hkdf` is the composition. It is `hmac` and `hashlib` and nothing else,
+  which is what puts it in btclib at all: issue #1066 settled BIP324's
+  transport line as "what can be built out of the standard library is in
+  scope, what would need a hand-rolled cipher is not", and HKDF-SHA256 is
+  the one piece of that key schedule on the near side of it. ChaCha20,
+  Poly1305, the AEAD, the forward-secure rekeying and the packet encoding
+  stay out, and so do BIP324's own labels, which are transport work with
+  no consumer here.
+
+  **`ansi_x9_63_kdf` moves with it**, out of `btclib.ecc.dh` and out of
+  `btclib.ecc`'s exports; RELEASE_NOTES.md has the spellings that breaks
+  and what to write instead. The module it was in is Diffie-Hellman, and
+  SEC 1 does put its KDF there -- section 3.6.1 for the agreement of
+  section 6.1 -- so with one KDF in the tree the arrangement cost
+  nothing. HKDF has no such tie: RFC 5869 has no Diffie-Hellman in it,
+  and nothing wanting HKDF here holds a public key, so leaving the drawer
+  under `ecc` meant either a KDF reached through a package of signature
+  schemes or two KDFs in two places for the next one to have to choose
+  between. A KDF is a hash construction with no curve and no bitcoin in
+  it, so the top level is the layer it belongs to, beside `hashes` --
+  which is what CONTRIBUTING.md's "Breaking a caller is not an argument"
+  asks to weigh this against, and the only cost is the note. `ecc.dh`
+  imports the module rather than the function, so
+  `btclib.ecc.dh.ansi_x9_63_kdf` is not a surviving spelling either: the
+  break is total rather than half-kept, which is what lets the release
+  note say so.
 
   RFC 5869's appendix A is the authority the tests use -- all seven
   vectors, SHA-256 and SHA-1, the zero-length salt, the absent salt and
   the zero-length info among them -- and it is third-party rather than
-  Bitcoin Core's, which is the kind of vector issue #993 asked for. Each
-  step is asserted against the published intermediate as well as the
-  composition against the final output: a `hkdf` with the two arguments
-  crossed over answers bytes of the right length, and only the prk says
-  which bytes.
+  Bitcoin Core's, which is the kind of vector issue #993 asked for. The
+  SHA-1 half is run rather than trimmed, and `tests/kdf_test.py`'s
+  docstring says why: this tree's rule is that btclib never *chooses*
+  SHA-1, and a KDF chooses no hash function at all -- `hf` is the
+  caller's, `ansi_x9_63_kdf` is already driven over `[sha1, sha256,
+  sha512]`, and running half of a published vector set is editing the
+  authority rather than citing it. Each step is asserted against the
+  published intermediate as well as the composition against the final
+  output: a `hkdf` with the two arguments crossed over answers bytes of
+  the right length, and only the prk says which bytes.
 
-  Held in `ecc/dh.py` beside `ansi_x9_63_kdf` rather than moved with it
-  into a KDF module of its own. RFC 5869 has no Diffie-Hellman in it, but
-  `ansi_x9_63_kdf` is SEC 1 section 3.6.1 and `diffie_hellman` is section
-  6.1 built on it -- its only caller here -- so a split would separate a
-  scheme from the derivation its own specification defines for it, and
-  `btclib.ecc` re-exports both KDFs anyway, so `from btclib.ecc import
-  hkdf` is the spelling either way. Nothing is source-breaking and
-  RELEASE_NOTES.md does not move. The size checks the two KDFs share are
-  now one `_assert_valid_keying_data_size`, the ceiling being the only
+  The size checks the two KDFs share are one
+  `_assert_valid_keying_data_size`, the ceiling being the only
   difference: SEC 1 numbers its blocks with four octets and RFC 5869 with
   one, so 255 digests is a length a test can ask for and measure where
   the other's ceiling is a hundred and thirty-seven gigabytes out of
